refactor(pcc-dashboard-service-management): move files
