+++ client/app/dedicatedCloud/dashboard/tiles/service-management/service-management.component.js
@@ -1,4 +1,4 @@
-import template from './serviceManagement.html';
+import template from './service-management.html';
 
 export default {
   bindings: {
