+++ client/app/dedicatedCloud/dashboard/tiles/service-management/service-management.html
